Namespace report classes under Reports::
