+++ app/mailers/reports_mailer.rb
@@ -10,7 +10,7 @@ class ReportsMailer < ApplicationMailer
 
   def employer_recruiter_daily_report
     @administrator = params[:administrator]
-    @sections = DailyReport.new(@administrator).sections
+    @sections = Reports::Daily.new(@administrator).sections
     @service_name = @administrator.organization.service_name
 
     mail to: @administrator.email, subject: t(".subject", service_name: @service_name)
@@ -18,7 +18,7 @@ class ReportsMailer < ApplicationMailer
 
   def employment_authority_weekly_report
     @administrator = params[:administrator]
-    @sections = WeeklyReport.new(@administrator).sections
+    @sections = Reports::Weekly.new(@administrator).sections
     @service_name = @administrator.organization.service_name
     @week_starts_on = 1.week.ago.beginning_of_week.to_date
