+++ config/locales/fr.yml
@@ -232,12 +232,13 @@ fr:
       job_applications: "Mes candidatures"
       job_offers: "Nos offres"
       logout: "Déconnexion"
-  daily_report:
-    sections:
-      new_offers: "Nouvelle offre"
-  weekly_report:
-    sections:
-      new_offers: "Nouvelle offre"
+  reports:
+    daily:
+      sections:
+        new_offers: "Nouvelle offre"
+    weekly:
+      sections:
+        new_offers: "Nouvelle offre"
   account:
     job_application_files:
       update:
